+++ src/config.rs
@@ -13,16 +13,16 @@ pub struct Config {
 }
 
 impl Config {
-    pub fn load() -> Result<Self, Box<dyn std::error::Error + Send + Sync>> {
+    pub fn load() -> Result<Self, Box<dyn std::error::Error>> {
         Ok(Self {
-            redis_addr: env::var("REDIS_ADDR")?,
-            socket_path: env::var("UNIX_SOCKET")?,
+            redis_addr: env::var("REDIS_ADDR").unwrap_or_else(|_| "redis://localhost:6379".to_string()),
+            socket_path: env::var("UNIX_SOCKET").unwrap_or_else(|_| "/tmp/hyperlocal.sock".to_string()),
             queue_name: "payments:queue".to_string(),
             sorted_set_default: "payments:list:default".to_string(),
             sorted_set_fallback: "payments:list:fallback".to_string(),
-            workers_len: env::var("WORKERS")?.parse::<u32>().unwrap_or(1),
-            payment_default_url: env::var("PAYMENT_PROCESSOR_URL_DEFAULT")?,
-            payment_fallback_url: env::var("PAYMENT_PROCESSOR_URL_FALLBACK")?,
+            workers_len: env::var("WORKERS").unwrap_or_else(|_| "1".to_string()).parse::<u32>().unwrap_or(1),
+            payment_default_url: env::var("PAYMENT_PROCESSOR_URL_DEFAULT").unwrap_or_else(|_| "http://localhost:8001".to_string()),
+            payment_fallback_url: env::var("PAYMENT_PROCESSOR_URL_FALLBACK").unwrap_or_else(|_| "http://localhost:8002".to_string()),
         })
     }
 }